[nandini] Refactored : Refcatr 4b - created a different jar for csv builder and incorporated it with statecensusanalyser class
