+++ src/main/java/CSVBuilder/ICSVBuilder.java
@@ -1,0 +1,11 @@
+package CSVBuilder;
+
+import java.io.Reader;
+import java.util.Iterator;
+
+import IndianStateCensus.StateCensusAnalyserException;
+
+public interface ICSVBuilder {
+	
+	public <E> Iterator<E> getCsvFileIterator(Reader reader, Class csvClass) throws StateCensusAnalyserException, StateCensusAnalyserException;
+}
