--- src/main/java/IndianStateCensus/ICSVBuilder.java
@@ -1,9 +1,0 @@
-package IndianStateCensus;
-
-import java.io.Reader;
-import java.util.Iterator;
-
-public interface ICSVBuilder {
-	
-	public <E> Iterator<E> getCsvFileIterator(Reader reader, Class csvClass) throws StateCensusAnalyserException;
-}
