+++ src/main/java/IndianStateCensus/StateCensusAnalyserException.java
@@ -3,7 +3,7 @@ package IndianStateCensus;
 //Exception class and types
 public class StateCensusAnalyserException extends Exception {
 
-	enum ExceptionType {
+	public enum ExceptionType {
 		INVALID_FILE_PATH,INVALID_DELIMITER,INVALID_HEADER,INVALID_TYPE,UNABLE_TO_PARSE;
 	}
